Move Queue.js to lib
