+++ src/queues/GambitCampaignSignupRelayQ.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const CampaignSignupMessage = require('../messages/CampaignSignupMessage');
-const Queue = require('./Queue');
+const Queue = require('../lib/Queue');
 
 class GambitCampaignSignupRelayQ extends Queue {
   constructor(...args) {
