+++ src/queues/GambitChatbotMdataQ.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const MdataMessage = require('../messages/MdataMessage');
-const Queue = require('./Queue');
+const Queue = require('../lib/Queue');
 
 class GambitChatbotMdataQ extends Queue {
   constructor(...args) {
